Make <*> lazy in its second argument where possible.
On `master`, use of `Control.Monad.forever` diverges. `forever` is defined as:

```haskell
forever a = let a' = a *> a' in a'
```

This definition of `forever` does not diverge:

```haskell
forever a = let a' = a >> a' in a'
```

It appears that `<*>` is a little too strict in its second argument. This PR resolves that by not forcing the evaluation of the second argument unless the first argument is in `Val`.