Implement a Resource effect and remove broken `bracket`.
The current formulation of `bracket` is broken in the presence of
asynchronous exceptions. If the action passed to `bracket` rethrows an
exception, and the exception is delivered asynchronously, there can
occur a race where the cleanup handler is invoked twice, which is
disastrous in terms of correctness.

The correct way to do this is to use `Control.Exception.bracket`, which
correctly masks asynchronous exceptions while executing the body of a
`bracket` call. We accomplish this by creating a new `Resource` effect
that uses `interpret` and a natural transformation to `IO` to invoke
the `bracket` function provided by `base`.