Add NFData instance for Resumable's SomeExc type.
If you're benchmarking the result of running an effect that supports
resumable exceptions, you'll want to have both sides of the resulting
`Either` implement `NFData` so that you can fully evaluate the result.

Unfortunately, `SomeExc` didn't implement `NFData`, so the `NFData`
instance for `Either (SomeExc exc) a` couldn't kick in. This patch
fixes that.

I'm not thrilled about adding another dependency, but `deepseq` is
ubiquitous in the ecosystem at this point.